server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-dotenv==1.0.1
redis==5.1.1
celery==5.4.0
agno==1.3.4
groq==0.11.0
resend==2.3.0
python-socketio==5.11.4
aiohttp==3.10.10
jinja2==3.1.4
pydantic-settings==2.5.2
reportlab==4.2.5
pandas==2.2.3
numpy==1.26.4
httpx==0.27.2
